+++ Assets/_Main/Scripts/GamePlay/DamageSystem/IDamageDealer.cs
@@ -1,5 +1,5 @@
 
 public interface IDamageDealer
 {
-    public void DealDamage(int damage, IDamagable damagable, DamageDealerType damageDealerType);
+    public void DealDamage(float damage, IDamagable damagable, DamageDealerType damageDealerType);
 }
